// richmindale-app.client/src/app/shared/models/admin/admission-details.model.ts
export class AdmissionDetails {
    Id: any;
    ApplicationRefNo: any;
    CreditTransfer: any;
    AdmissionTypeId: any;
    RequestTypeId: any;
    CountryId: any;
    CountryName:any;
    LrnCampusId: any;
    CampusName:any;
    LearningMethodsId: any;
    LearningmethodName:any;
    ProgramCategoryId: any;
    ProgramCategoryName: any;
    ProgramId: any;
    ProgramTitle: any;
    ProgramCode: any;
    TermId: any;
    TermName: any;
    EmailAddress: any;
    ApplyDiscount: any;
    DiscountTypeId: any;
    DiscountTypeName: any;
    DiscountCode: any;
    Status: any;
    StatusName:any;
    // Student
    StudentCode:any;
    StudentLastname: any;
    StudentGivenName: any;
    StudentMiddleName: any;
    StudentMaidenName: any;
    StudentBirthDateLong: any;
    StudentBirthDate: any;
    StudentGender: any;
    StudentGenderName: any;
    StudentMaritalStatus: any;
    StudentMaritalStatusName:any;
    StudentNationality: any;
    StudentNationalityName:any;
    StudentReligion: any;
    StudentReligionName:any;
    StudentCountry: any;
    StudentCountryName:any;
    StudentState: any;
    StudentStateName:any;
    StudentCity: any;
    StudentCityName: any;
    StudentZipCode: any;
    StudentAddress: any;
    StudentPhoneCode1: any;
    StudentPhone1: any;
    StudentEmail1: any;
    StudentPhoneCode2: any;
    StudentPhone2: any;
    StudentEmail2: any;
    // Previous School
    SkipSchool: any;
    SchoolName: any;
    Qualifications: any;
    QualificationName:any;
    SchoolProgram: any;
    SchoolProgramName:any;
    SchoolOtherProgram:any;
    SchoolEmail: any;
    SchoolCountry: any;
    SchoolCountryName:any;
    SchoolAddress: any;
    SchoolContact: any;
    SchoolContactDesignation: any;
    SchoolContactEmail: any;
    SchoolPhoneCode: any;
    SchoolContactNumber: any;
    // Father 
    SkipFather: any;
    FatherLastname: any;
    FatherGivenName: any;
    FatherMiddleName: any;
    FatherMaidenName: any;
    FatherBirthDate: any;
    FatherGender: any;
    FatherMaritalStatus: any;
    FatherMaritalStatusName:any;
    FatherNationality: any;
    FatherNationalityName:any;
    FatherReligion: any;
    FatherReligionName:any;
    FatherCountry: any;
    FatherCountryName: any;
    FatherState: any;
    FatherStateName:any;
    FatherCity: any;
    FatherCityName:any;
    FatherZipCode: any;
    FatherAddress: any;
    FatherPhoneCode1: any;
    FatherPhone1: any;
    FatherEmail1: any;
    FatherPhoneCode2: any;
    FatherPhone2: any;
    FatherEmail2: any;
    // Mother
    SkipMother: any;
    MotherLastname: any;
    MotherGivenName: any;
    MotherMiddleName: any;
    MotherMaidenName: any;
    MotherBirthDate: any;
    MotherGender: any;
    MotherMaritalStatus: any;
    MotherMaritalStatusName: any;
    MotherNationality: any;
    MotherNationalityName:any;
    MotherReligion: any;
    MotherReligionName:any;
    MotherCountry: any;
    MotherCountryName:any;
    MotherState: any;
    MotherStateName:any;
    MotherCity: any;
    MotherCityName: any;
    MotherZipCode: any;
    MotherAddress: any;
    MotherPhoneCode1: any;
    MotherPhone1: any;
    MotherEmail1: any;
    MotherPhoneCode2: any;
    MotherPhone2: any;
    MotherEmail2: any;
    // Guardian
    SkipGuardian: any;
    GuardianLastname: any;
    GuardianGivenName: any;
    GuardianMiddleName: any;
    GuardianMaidenName: any;
    GuardianBirthDate: any;
    GuardianGender: any;
    GuardianMaritalStatus: any;
    GuardianMaritalStatusName:any;
    GuardianNationality: any;
    GuardianNatinalityName:any;
    GuardianReligion: any;
    GuardianReligionName:any;
    GuardianCountry: any;
    GuardianCountryName:any;
    GuardianState: any;
    GuardianStateName:any;
    GuardianCity: any;
    GuardianCityName:any;
    GuardianZipCode: any;
    GuardianAddress: any;
    GuardianPhoneCode1: any;
    GuardianPhone1: any;
    GuardianEmail1: any;
    GuardianPhoneCode2: any;
    GuardianPhone2: any;
    GuardianEmail2: any;
}